summary/summarization/requirements.txt
transformers>=4.30.0
torch>=1.10.0
supabase>=2.0.0
python-dotenv>=0.19.0
tqdm>=4.60.0
sentencepiece>=0.1.96
pillow>=9.0.0